+++ src/app/FormControls.js
@@ -1,151 +1,109 @@
 import React from 'react';
-import { RIEInput, RIEToggle, RIETextArea, RIENumber, RIESelect } from 'riek';
-import _ from 'lodash';
 
-export function TextInput({ value, propName, label, change, disabled=true }) {
+export function TextInput({ value, prop, label, change }) {
   return (
     <label className="label">
       {label}
-      <RIEInput
-        isDisabled={disabled}
-        value={value || ''}
-        change={change}
-        propName={propName}
+      <input
         className="input"
-        validate={_.isString} />
-    </label>
-  );
-}
-
-export function TextArea({ value, propName, label, change, disabled=true }) {
-  return (
-    <label className="label">
-      {label}
-      <RIETextArea
-        isDisabled={disabled}
         value={value || ''}
-        change={change}
-        propName={propName}
-        rows={8}
-        columns={50}
-        className="textarea"
-        validate={_.isString} />
-    </label>
-  );
-}
-  
-export function NumberInput({ value, propName, label, change, disabled=true }) {
-  return (
-    <label className="label">
-      {label}
-      <RIENumber
-        isDisabled={disabled}
-        value={value || 0}
-        change={change}
-        propName={propName}
-        className="input"
-        validate={_.isNumber} />
-    </label>
-  );
-}
-
-export function DateInput({ value, propName, label, change, disabled=true }) {
-  return (
-    <label className="label">
-      {label}
-      <RIEInput
-        isDisabled={disabled}
+        name={prop}
         type="text"
-        value={value || ''}
-        change={change}
-        propName={propName}
-        className="input"
-        validate={_.isDate} />
+        placeholder={label}
+        onChange={change} />
     </label>
   );
 }
 
-export function Toggle({ value, propName, label, change, disabled=true }) {
+export function UrlInput({ value, prop, label, change }) {
   return (
     <label className="label">
       {label}
-      <RIEToggle
-        isDisabled={disabled}
-        value={value || 'No'}
-        change={change}
-        propName={propName}
-        textTrue="Yes"
-        textFalse="No"
+      <input
         className="input"
-        validate={_.isString} />
+        value={value || ''}
+        name={prop}
+        type="text"
+        placeholder={label}
+        onChange={change} />
     </label>
   );
 }
 
-export function ToggleEditMode({ value, propName, change, disabled=true }) {
-  // TODO: need to work on this
-
-  return (
-    <div>
-      <icon className="fa fa-check fa-lg" />
-      <icon className="fa fa-pencil fa-lg" />
-    </div>
-  //   <RIEToggle
-  //     isDisabled={disabled}
-  //     value={value}
-  //     change={change}
-  //     propName={propName}
-  //     textTrue={pencil}
-  //     textFalse={check}
-  //     className="select" />
-  );
-}
-
-const fakeOptions = [{ id: -1, text:'' }];
-
-export function TextSelect({ value, propName, label, options=fakeOptions, change, disabled=true }) {
+export function NumberInput({ value, prop, label, change }) {
   return (
     <label className="label">
       {label}
-      <RIESelect
-        isDisabled={disabled}
-        value={value || ''}
-        change={change}
-        propName={propName}
-        options={options}
+      <input
         className="input"
-        validate={_.isString} />
+        value={value || 0}
+        name={prop}
+        type="number"
+        placeholder={label}
+        onChange={change} />
     </label>
   );
 }
 
-export function UrlInput({ value, propName, label, change, disabled=true }) {
+export function Checkbox({ value, prop, label, change }) {
   return (
-    <label className="label">
-      {label}
-      <RIEInput
-        isDisabled={disabled}
-        value={value || ''}
-        change={change}
-        propName={propName}
-        className="input"
-        validate={_.isString} />
+    <label className="checkbox">
+      <input
+        className="checkbox"
+        value={value || false}
+        name={prop}
+        type="checkbox"
+        onChange={change} />
+      <span>{label}</span>
     </label>
   );
 }
 
-export function EmailInput({ value, propName, label, change, disabled=true }) {
+export function DateInput({ value, prop, label, change }) {
   return (
     <label className="label">
       {label}
-      <RIEInput
-        isDisabled={disabled}
-        type="email"
+      <input
+        className="input"
+        value={value || '1987-04-23' }
+        name={prop}
+        type="date"
+        placeholder={label}
+        onChange={change} />
+    </label>
+  );
+}
+
+export function Dropdown({ value, prop, label, options, change }) {
+  return (
+    <label className="label">
+      {label}
+      <select
+        className="input"
+        value={value || options[0]}
+        name={prop}
+        onChange={change}>
+
+        {options.map((text, index) => {
+          return <option key={index} value={text}>{text}</option>;
+        })}
+
+      </select>
+    </label>
+  );
+}
+
+export function TextArea({ value, prop, label, change }) {
+  return (
+    <label className="label">
+      {label}
+      <textarea
+        className="input"
         value={value || ''}
-        change={change}
-        propName={propName}
-        className="input has-icons-left"
-        validate={_.isString} />
+        name={prop}
+        rows="10"
+        cols="50"
+        onChange={change} />
     </label>
   );
 }
