+++ src/app/GlobalHeader.js
@@ -11,7 +11,14 @@ export default function GlobalHeader({ authId, signOut }) {
               <Link to="/"><i className="fa fa-home fa-2x"></i></Link>
             </li>
             <li>
-              <input className="input" placeholder="search" />
+              {/* <input className="input" placeholder="search" /> */}
+              <div className="control has-icons-left">
+                <input className="input" placeholder="Search" />
+                <span className="icon is-small is-left">
+                  <i className="fa fa-search"></i>
+                </span>
+              </div>
+
             </li>
             <li>
               <Link to="/about">About</Link>
