+++ src/info/InfoContainer.js
@@ -84,6 +84,7 @@ export class InfoContainer extends Component {
   }
 
   render() {
+    this.updateInitialState();
     return (
       <div>
         {this.state.editAllowed ? <ToggleEditor editModeOn={this.state.editModeOn} toggleFn={this.toggleEditMode} /> : null }
