+++ src/media/MediaGallery.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 // import 'bulma/css/bulma.css';
 // import PropTypes from 'prop-types';
-import { TextSelect, TextArea, UrlInput, Toggle, ToggleEditMode } from '../app/FormControls';
+import { TextSelect, TextArea, UrlInput, Toggle, ToggleEditMode } from '../app/FormControlsRIEK';
 import { updateMedia, getMedia } from './actions';
 
 // GalleryItem.propTypes = {
